Added function for parent to delete their own account
Updated the Users and ParentReviews table so that when a user is deleted, all related reviews from the user is deleted too.

all kids account created by the parent account is deleted too when the parent deletes their account.

Even if the review is being showcased by the admin, it will be deleted too